requirements.txt
google-generativeai
openai
python-dotenv
telegram
schedule
pyjwt
cryptography
requests